Refactor to use traits and extend base collection instead of eloquent collection
